+++ src/Aggregation/GeohashGrid.php
@@ -39,12 +39,16 @@ class GeohashGrid extends AbstractAggregation
     /**
      * Set the precision for this aggregation.
      *
-     * @param int $precision an integer between 1 and 12, inclusive. Defaults to 5.
+     * @param int|string $precision an integer between 1 and 12, inclusive. Defaults to 5 or distance like 1km, 10m
      *
      * @return $this
      */
-    public function setPrecision(int $precision): self
+    public function setPrecision($precision): self
     {
+        if (!\is_int($precision) && !\is_string($precision)) {
+            throw new \TypeError(\sprintf('Argument 1 passed to "%s()" must be of type int|string, %s given.', __METHOD__, \is_object($precision) ? \get_class($precision) : \gettype($precision)));
+        }
+
         return $this->setParam('precision', $precision);
     }