+++ tests/Aggregation/ExtendedStatsTest.php
@@ -32,6 +32,27 @@ class ExtendedStatsTest extends BaseAggregationTest
         $this->assertArrayHasKey('sum_of_squares', $results);
     }
 
+    /**
+     * @group functional
+     */
+    public function testExtendedStatsAggregationWithMissing(): void
+    {
+        $agg = new ExtendedStats('stats');
+        $agg->setField('price');
+        $agg->setMissing(10);
+
+        $query = new Query();
+        $query->addAggregation($agg);
+        $results = $this->_getIndexForTest()->search($query)->getAggregation('stats');
+
+        $this->assertEquals(5, $results['count']);
+        $this->assertEquals(1, $results['min']);
+        $this->assertEquals(10, $results['max']);
+        $this->assertEquals((5 + 8 + 1 + 3 + 10) / 5.0, $results['avg']);
+        $this->assertEquals((5 + 8 + 1 + 3 + 10), $results['sum']);
+        $this->assertArrayHasKey('sum_of_squares', $results);
+    }
+
     protected function _getIndexForTest(): Index
     {
         $index = $this->_createIndex();
@@ -41,6 +62,7 @@ class ExtendedStatsTest extends BaseAggregationTest
             new Document(2, ['price' => 8]),
             new Document(3, ['price' => 1]),
             new Document(4, ['price' => 3]),
+            new Document(5, ['anything' => 'anything']),
         ]);
 
         $index->refresh();